astro_api/requirements.txt
Django==3.1.7
asgiref==3.3.1
djangorestframework==3.12.4
pip==21.0.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
setuptools==54.2.0
sqlparse==0.4.1